requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.0
aiosignal==1.3.2
antlr4-python3-runtime==4.9.3
asttokens==3.0.0
attrs==25.3.0
decorator==5.2.1
executing==2.2.0
filelock==3.18.0
frozenlist==1.6.0
fsspec==2025.5.1
graphviz==0.20.3
hydra-core==1.3.2
idna==3.10
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
lightning==2.5.1.post0
lightning-utilities==0.14.3
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.4.4
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
omegaconf==2.3.0
optree==0.15.0
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
prompt_toolkit==3.0.51
propcache==0.3.1
ptyprocess==0.7.0
pure_eval==0.2.3
pybind11==2.13.6
Pygments==2.19.1
pytorch-lightning==2.5.1.post0
PyYAML==6.0.2
scipy==1.15.3
setuptools==80.8.0
stack-data==0.6.3
sympy==1.14.0
torch==2.7.0
torchmetrics==1.7.1
torchopt==0.7.3
torchvision==0.22.0
torchviz==0.0.3
tqdm==4.67.1
traitlets==5.14.3
triton==3.3.0
typing_extensions==4.13.2
wcwidth==0.2.13
yarl==1.20.0
